+++ app/components/buttons/Filter.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="flex gap-1 items-center border-2 border-accent rounded-full px-2">
+    <p class="font-inter text-accent">no filter</p>
+    <Icon name="mdi:close" size="16" class="text-accent" />
+  </div>
+</template>
+
+<script lang="ts" setup>
+
+</script>
+
+<style>
+
+</style>